[#15] feat : 좌석 선택 후 시간 선택하지 않으면 결제페이 이동하지 못하게 하는 기능 구현
