+++ catch-study/src/components/BookingSeatModal/BookingSeatModal.tsx
@@ -1,37 +1,23 @@
-import { useEffect, useState } from 'react';
 import { SeatPriceTypes } from '../../types/interfaces';
 
 interface BookingSeatModalPropTypes {
   usageFee: SeatPriceTypes[];
-  selectedSeat: {
-    type: string;
-    id: number;
-  };
+  priceClick: (
+    e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
+    price: number,
+  ) => void;
+  selectedPrice: number;
 }
 
 const BookingSeatModal: React.FC<BookingSeatModalPropTypes> = ({
   usageFee,
-  selectedSeat,
+  priceClick,
+  selectedPrice,
 }) => {
-  const [selectedPrice, setSelectedPrice] = useState(0);
-
-  const priceClick = (
-    e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-    price: number,
-  ) => {
-    e.preventDefault();
-
-    setSelectedPrice(price);
-  };
-
-  useEffect(() => {
-    setSelectedPrice(0);
-  }, [selectedSeat]);
-
   return (
     <>
       {usageFee.map((fee, i) => (
-        <div key={i} className='w-1/2 h-80 p-10'>
+        <div key={i} className='w-1/2 p-10 h-80'>
           <button
             onClick={e => priceClick(e, fee.price)}
             className={`w-full h-full border-dark-gray rounded-sm text-16 ${selectedPrice === fee.price ? 'bg-blue text-white' : 'border-[1px]'}`}
